+++ src/ValueObject/Point.php
@@ -2,7 +2,10 @@
 
 namespace VysokeSkoly\ImageApi\Sdk\ValueObject;
 
-class Point
+use Imagine\Image\BoxInterface;
+use Imagine\Image\PointInterface;
+
+class Point implements PointInterface
 {
     public function __construct(private readonly int $x, private readonly int $y)
     {
@@ -16,5 +19,20 @@ class Point
     public function getY(): int
     {
         return $this->y;
+    }
+
+    public function in(BoxInterface $box): bool
+    {
+        return $this->x < $box->getWidth() && $this->y < $box->getHeight();
+    }
+
+    public function move($amount): self
+    {
+        return new self($this->x + $amount, $this->y + $amount);
+    }
+
+    public function __toString(): string
+    {
+        return sprintf('(%d, %d)', $this->x, $this->y);
     }
 }